requirements.txt
# Core Dependencies
langgraph>=0.0.40
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0

# MCP Protocol
mcp>=0.1.0
httpx>=0.25.0

# Data Processing
pandas>=2.0.0
python-dotenv>=1.0.0

# Output Generation
markdown>=3.5
reportlab>=4.0.0

# Utilities
pydantic>=2.0.0
asyncio

# Web UI
streamlit>=1.30.0
